gpu: Offset ray to prevent self-intersection
Based on the article from Ray Tracing Gems book (chapter 6).

The part in the book which explains special case around origin is IMHO a bit unclear
and below is my attempt to clarify this.

They are trying to say that in near-zero region numeric error is mostly due to
direction components - position components are small and error due to
direction components starts to dominate. That's also why the error in near-zero
region is almost constant - error contribution from components of constant-length
direction vector on average is the same. We can use constant epsilon to cover this error.

Position is different - the same as for direction the error magnitude is proportional
to vector length. But position is not a unit vector like direction. Error increases with
the magnitude of position components.