Update coordinate system handling code
The new approach is to not have some program's preferable
coordinate system and to work in the coordinates of the input
scene. We mostly do not care about the axis direction, there is
functionality, like camera vertical direction movement, which require
knowledge of the _up_ axis. We also check for handedness in
few places - for example, PBRT has left-handed conventions and
we ensure that output matches PBRT's output by handling this in
small part of the program but most code should not know  about
this and should assume right-handed conventions.