ref: Fix for infinite pdf (infinity pdf is ok!)
Yes, infinite pdf is not an issue, it happens in rare situations
for valid reasons (for example in env map sampler when random
variable happens to be u = 0.0) and in that case the radiance contribution
is computed as L = (Le * f * cos) / pdf = X / inf = 0, when X != 0 according to ieee 754.
The problem with pdf == inf is when we have MIS weight L = weight * (Le * f * cos) / pdf.
weight = pdf1 / (pdf1 + pdf2) and inf/inf = NaN. It's easy to show that for this
specific case the expected weight for light sampling should be 1.0, and corresponding bsdf
mis weight will be computed as bsdf_pdf/Inf = 0.0 which is also correct.