ref: Fix another bug that zero radiance samples are ignored
When we compute integral all samples have the same rights, zero value sample it's just
another value which affects the resulted estimate of the integral.

Example where previous implementation gave incorrect resut: through one part of the
given pixel we can see area light source and the rest part of the pixel does not project
into the scene (it's vacuum). The average radiance for this pixel should not be as bright
as the light source because the part that contains vacuum should decrease the average
brightness.