ref: Handle zeroed rows in Distribution_2D
If row contains all zeros it should never be sampled and as such,
we don't try to compute cdf for that row. If that row is selected
due to bug the program witll assert due to invalid cdf.

Also added handling of special case when CDF function starts
with horizontal segment.