ref: Update trace_ray implementation
Ray and auxilary rays are input only parameters now.
If there is a hit then we don't need to access the final ray direction
in case it was changed by a specular surface. If we don't have a hit
then the last traced ray is stored as Shading_Context::miss_ray.