ref: Fix premature mip pyramid lod value clamping
This led to failure of the entire filtering algorithm. For large UV derivatives (for example
after uv scaling is applied + uv was modified by specular reflection) we calculated lod which was
quite large (18 for example) and clamped it to max lod level (say 10) before passing it to EWA filtering
routine. The problem with this approach is that we still pass huge UV derivatives values to EWA filtering
algorithm and this results in large pixel area to iterate (might be million of pixels) which leads to very
long computation time. In addition large UV derivative values produce totally inaccurate floating point results
(especially when multiplied with each other) and one situation we had is that conic section expression A*C - 0.25*B*B,
which should be positive in our case sinse we are building an ellipse, were evaluated to negative value after
which the computations did not make sense.
So instead of clamping lod value we check explicitly if lod is bigger than mip pyramid size (it means
we have large UV derivatives) and in that case we return the value from the top of the pyramid.