ref: Make Specular_Scattering_Params part of Shading_Context
The 'specular_attenuation' member was removed from Shading_Context
because it's not the information that is associated with a single intersection
event but rather it aggregates contribution from multiple specular bounces.
This specular attenuation value is defined now only as part of higher level
functionality (returned by trace_specular_bounces function).