ref: Fix kdtree tests issue with intersection behind the ray origin
It was mostly done to reduce probability of the situation when a ray does not
intersects bounding box but the intersection with a triangle inside that bounding
box actually happens due to floating point rounding errors: intersection distance
't' along the ray is computed as a small value, like order of 1e-7. The real intersection
happens behind the ray origin but rounding errors drags us into a positive range.

One solution to fix the above issue is to do error estimation by doing numerical
analysis but measurement showed that it increases intersection time roughly 2x
and actually the main reason why we had 99.99% of that issue is because of how
kdtree test algorithm was coded.

Anyway here's the version of triangle intersection algorithm with numerical analysis
of rounding error (it's not included in the final solution):
https://gist.github.com/kennyalive/c3cc15e2208fdca8a72b9fce4b9874a8
It's a bit different from pbrt mostly because I did analysis myself and it happens
sometimes that different people come up with different results.