ref: Initialize Shading_Contex when no intersection is reported
Even when no intersection is found the Shading_Context is still used in
in the code with expectation that it is initialized with default values (except
the miss ray field). The bug was that we had perfect specular scattering state
from the previous sample and it was different from default specular scattering
state which is expected for miss intersection.

Additional safety measure is added: always reset shading context to default
state when start new sample (it's not stricly needed for current code which
always uses shading context after trace_ray call).