lib: Detect when vector length computation results in infinity
This happens for vectors with large components (~1e18). When squared it
turns into infinity and usually infinity lenght is not something computation
code expects as a valid result. If length infinity is something that works with
particular algorithm then length_squared() should not be used and instead
that value should be computed manually  as x*x+y*y+z*z or using dot(v, v) expression.