ref: Fix calculation of interval index + fix black tail rows sampling issue
The problem was related to FP rounding error. If we multiply value A
from  [0..1) by integer K then A * K can result exactly in K even if A is strictly
less than one. This can happen if rounding finds that 1.0f as the closest FP
value to 'exact answer'. We had a code that assumed A*K < K and that
led to incorrect results in a rare case of equality.

The solution is instead of trying to perform FP calculations and handle
corner cases we can actually return result (interval_index) directly.
The distribution sampling algorithm computes interval index as intermediate
variable and we can return it directly as additional output parameter.

This commit fixes one more bug. If the image ends with completely black rows
the sampling of marginal Y CDF might select such black row for subsequent X
sampling and that triggers assert (X CDF for completely black row is all 0s which
is invalid CDF and is detected by the code - such CDF should never be sampled).
The fix is to fill the elements of marginal Y CDF table that corresponds to black tail
rows with max CDF value 1.0, in this case they are never selected for sampling.

One more minor contribution of this commit. Some optional output parameters I
changed to be non-optional. The main reason is not to consume additional branch
predictor entries and to get rid of comparison+jump instructions.
For optional output parameters we have to write a condition. It's not a performance
problem per se because condition can be easily predicted in our use case but it also
consumes a predictor slot, so why not to use it for conditions we really need.
Plus we removed 'if' which simplifies the code and non-optional parameter simplifies
the interface.