ref: Make KdTree type non-parameterized
This simplifies code, makes core concepts more well defined.
Also it makes the code more extensible, this work is preparation before
adding support for testing against alpha texture during ray traversal.
Surprisingly, performance wise new code is often a bit faster even if
we added dynamic dispatch - simplicity wins! In some cases it might
be a bit slower but the difference is small, performance is not the
main gain of this change.