ref: Introduce Triangle_Mesh_Geometry_Data
Keep reference to this structure instead of to triangle mesh directly.
For transparency testing we need to put additional data alongside
triangle mesh reference.

One concern that additional level of memory indirection can have
measurable performance effect. I did not detect it but if it happens in
the future there is a solution to this - support two methods of setting
triangle mesh data - either to set Triangle_Mesh reference directly as
in the previous implementation or to set Triangle_Mesh_Geometry_Data
if additional data is required.