ref: Fix non-determinism when primitives overlap at intersection point
This fixes non-deterministic behavior with regard to using different kdtree versions
(i.e. built with different parameters). But with regard to specific setup of acceleration
structure the rendering is always deterministic (so multiple renders produce the same output).

There are other sources of non-deterministic behavior when using different versions
of kdtree. This might be due to known fundamental issue that a ray that misses bounding
box can still report intersection with a primitive inside it (due to rounding error). Also there
might be other reasons I did not think about. This can be reproduced with pbrt landscape-view0 scene.