ref: Remove handling of special case when ray is aligned with kdtree splitting plane
The original intention was ensure that kdtree result matches brute
force triangle intersection but that's not practical. The deviation
from brute force intersection is rather theoretical because even
different implementations of brute force intersection might produce
different results (for example, how to resolve multiple intersection at
the same position). The code that handles that case is ununtiutive
with very questionable benefits.

There is use case when the ray that is in the splitting plane goes through
the edges of multiple triangles which are located at different distances
along the ray. The original code tried to select the closest triangle. In practise
it does not important if we have multiple samples per pixel. If there are only
few samples it might be that in some cases the edge of further triangle is
selected as intersection - we qualify such result as one of many artifacts
associated for low sample count. With increased number of samples the
edge is resolved as expected.

Initially I tried to ensure that kdtree results stricly match brute force
intersection and implemented few fixes similar to the removed one. But
they ensure equivalence only for specific cases and not in the general
case. It's not easy to ensure equivalence due to rounding error in floating
point computations. It could be that bounding box intersection test reports
no intersection so we don't try to check primitives inside that node. But even
if triangle is completely inside bounding box the intersection test with that
triangle might report a hit, even there is no hit with enclosing bounding box!
All this is due to rounding errors - triangle intersection algorithm might
have bigger rounding error, so a triangle might 'stick out' of the bounding
box from intersection algorithm standpoint.

The discussed above mistmatch between kdtree result and brute force
triangle intersection does not mean that rendering results are wrong. It's
similar to the situation when we use two different algorithms for triangle
intersection which produce a bit different results due to finite precision
arithmetic.