ref: Add support for 'list' files (for batch rendering)
A list file contains the list of the file paths. Each file path should
be on a separate line. A list file is one more way to render
multiple scenes in a single run in addition to existed option
to specify each file path as a seperate command line argument.
List files might be useful for testing purposes when we want
to render a lot of scenes. The list files can also group scenes
based on some traits.