ref: New implementation of tiled layout (varying size nodes, no separate leaf/empty nodes)
10-20% memory reduction, 10-20% performance loss.
Initial expectations were that this might increase performance because we
don't loose memory inside tile due to varying size nodes so we can get
more traversal without caches misses and also we don't need to jump
when child is leaf/empty because that information is embedded in the
parent node. This approach led to some reduction of memory  usage
but supposedly more complicated traversal code slowed things a bit and
better nodes locality was not enough to compensate.

Going to profile this code a bit and check assemlby as a post-mortem.