ref: Handle non-enclosed trasmissive object
Nested dielectrics tracking is  disabled by default now.
It should be enabled per-object when needed. Often
transmisive objects in the scene are not real volumes
and we can not use nested dielectric tracking with them.
When nested dielectrics tracking is disabled we use surface
normal to determine outside/inside spaces.