ref: Enable openexr varying attributes by default
Previously they were disabled by default and the idea was it
will help to verify rendering determinism. Still it looks that
during development it's more useful to have all information
available and if you need deterministic output then this intention
should be specified explicitly as a command line flag.

Also determinism is easy to check even with varying attributes
stored in the openexr file. These attributes occupy a range of sequential
bytes, so it's easy to check if the difference is only that specific region.