ref: Apply russian roulette after new path segment is generated
Previously we applied it to the previous path state with intention
to save time on generating new path segment - it might be discarded
if russian roulette makes a decision to stop. That logic was probably
flawed. The previous path coeffs have larger values comparing to the
incremented path coeffs (due to energy lose during scattering), so
previous path state generates higher probability to continue tracing
even if newly generated segment creates a path with low contribution.