ref: Implement proper intersection position adjustment
Position adjustment is postponed until we have all the information. Just having
BSDF is not enough, we need to actually sample it to know in which hemisphere
we are working.

Previously we tried to offset point immediately during shading_ctx initialization.
Unfortunately, at that point, in the general case, we don't know the direction of adjustment.
It depends on the BSDF sampling result. Also because we use MIS, we might
need to offset position in both directions (along and against normal direction)
because we sample 2 directions (BSDF + Light sampling).