ref: Rect light: remove early exit optimization
The problem with that optimization that it did not take
into account that surface position will be adjusted later to
avoid self intersection. Even if the early condition determines
that a point is on the emitting side of the light, it might be not
the case after position adjustment. In those cases we need
to support special handling of infinite pdf by MIS heuristic
(we still need to do this for env lights but maybe in the future
we can fix that too) and also we had that paranoid asserts to
ensure that position adjustment does not adjust too much...

With new implementation we moved early exit a bit further,
so it's not an early exit anymore (but still not super late), to the
point where we can test light plane against adjusted position.
Surprisingly, on the ganesh scene this added measurable
performance improvement.. Friends, don't marry your
optimizations.