ref: Fix issues with roulette activation for delta scattering event
Delta scattering contribution was applied on the next loop iteration because
we re-used the ray cast at the beginning of the next iteration and it resulted
in a nice code with only a single ray cast for the entire algorithm (but not
much of other benefits).

Unfortunately it resulted in a probelm that russian roulette activated more
early for the delta events comparing to bsdf events. Additionally, the roulette
compensation coefficient 1/(1-q) was applied to the path before the light
contribution is computed on the next loop iteration. Lighting should always
be applied to the path value before it is adjusted by the roulette - the adjustment
ensures unbiased results for the subsequent paths but not for the current one.

The new code fixes the problem by handling delta scattering in the same way
as bsdf scattering - all computations are performed in the same loop
iteration.