ref: Add --openexr-compress to enable OpenEXR image compression
By default compression is disabled. Previously it was enabled by default,
but now I'm thinking that uncompressed image data is more convenient
for development, it's easy to detect changes in individual pixels. With
compression enabled the change in a single pixel might change large
chunk of binary data, which makes it hard to estimate the amount of
changed pixels.