ref: Implement rendering resume feature
This feature allows to close console window or to terminate program's process
at any moment and in any (brutal) way without losing current  rendering
progress.  When you start rendering again it will continue from where it left.

To enable this feature specify --checkpoint <directory_path> on the command
line with the path of an empty or non-existent directory that will be used to
store rendering progress. Then, when resuming rendering session specify the
same parameters to continue from that checkpoint.

The resume granularity is based on the film's tile size. The progress for all
in-progress tiles is lost when the renderer process terminates but they will be
re-rendered during the next session. There is also a guarantee that all the tiles
stored on disk are fully finished and that the writes are done is an atomic way
(by using atomic rename), so even if the process terminates during write
operation, it just means that the tile won't be added to the checkpoint and it
will be rendered again on the next session.

Total rendering time/variance statistic does not work yet when the checkpoint
rendering is active. This will be fixed in the next few commits.