ref: Replace enkiTS with custom threading solution
It's simpler and runs at the same speed.

Actually new code might be a bit faster but I doubt the performance
improvements are measurable. The rendering part is much heavier
then scheduling part, and it's not important if we use enkiTS or
this new solution or even some heavy solution. Still scheduling
part of current solution is just increment of atomic variable, that's
all. It's very simple. Also we get rid of enkiTS dependency.