ref: Fix bug when thin non-empty leaf node is skipped during traversal
This was reproducible in Sanmiguel pbrt3 scene. Separate regression test was added
to check this scenario.

The problem is when we have thin node (like few ulps) but not completely flat, then
mathematically we might expect, for example, that t_min is different from t_split,
because the ray has to traverse some distance from t_min to the splitting plane, and
obviously we can't skip the node we enter at t_min and exit at the splitting plane.

But for thin node the difference between t_min and t_split (computed with infinite
precision) might be smaller than float precision, so t_min and t_split will evaluate to
the same value. In that case in the old code we made decision to select only one
node because of the equality part in t_split >= t_max or t_split <= t_min expressions.
The correct way to think about this is that equality t_split == t_max or t_split == t_min
might indicate a situation described above and it's not safe to skip one of the nodes
and we have to traverse them both.