ref: Add pbrt3 version of plastic
Pbrt3 has a known bug with plastic, and because pbrt3 is finished and is not updated
that bug now is a feature, and we need a way to replicate this behavior when we need
the same output as in pbrt3.

We also have our version of plastic which behaves properly and it should be used in the
general case.