ref: Add pbrt3 fourier bsdf
This implementation supports only basic features to produce similar
result as pbrt3. It does not support interpolation and important
sampling in the form that pbrt3 does (we use basic cosine sampling).

For the pavilion scene the results are similar to pbrt3 output. If other
scenes produce more divergence, then one option is to add interpolation.

Because this bsdf is not supported in pbrt4, there is no enough stimulus
to match all pbrt3 features. The plan is to support what is needed to
produce good renders of existing pbrt3 scenes that use fourier material.