ref: Check for zero pdf in rough glass bsdf
Due to fp rounding errors it's possible to have a minor mismatch
between evaluate() and pdf() computations. pdf() can return 0.0 but
evaluate() might result in a very small but non-zero value. If pdf is
zero, then return Black radiance value because that's how the caller
checks that scaterring is not happening (it does not check pdf).