ref: Fix rought glass bsdf computation
The intuition that there should be outgoing IoR in the numerator.
If you trasmit into more dense environment (higher IoR) then the
radiance increases, so the larger value IoR value (output) should be in
the numerator. The same logic works if transmission happens into the
less dense environment. I also made a precise derivation that gets
this result.